pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "market-digest-discord-bot"
version = "0.1.0"
description = "Discord bot that posts a daily macro, stock, and crypto news digest."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.5,<3",
  "google-genai>=1.0,<2",
  "httpx>=0.28,<0.29",
  "python-dotenv>=1.0,<2",
  "tzdata>=2025.2",
]

[project.scripts]
market-digest-bot = "market_digest_bot.bot:run"

[tool.hatch.build.targets.wheel]
packages = ["src/market_digest_bot"]
